+++ Clockwork/garrysmod/gamemodes/clockwork/framework/config/sv_config.lua
@@ -111,3 +111,4 @@ Clockwork.config:Add("chat_multiplier", true, true, true);
 Clockwork.config:Add("steam_api_key", "");
 Clockwork.config:Add("enable_map_props_physgrab", false);
 Clockwork.config:Add("translate_api_key", "");
+Clockwork.config:Add("entity_handle_time", 0.1);