+++ Clockwork/garrysmod/gamemodes/clockwork/plugins/doorcommands/plugin/commands/sh_doorsetchild.lua
@@ -17,11 +17,20 @@ function COMMAND:OnRun(player, arguments)
 	
 	if (IsValid(door) and Clockwork.entity:IsDoor(door)) then
 		if (IsValid(player.cwParentDoor)) then
-			cwDoorCmds.parentData[door] = player.cwParentDoor;
-			cwDoorCmds:SaveParentData();
-			
-			Clockwork.entity:SetDoorParent(door, player.cwParentDoor);
-			Clockwork.player:Notify(player, "You have added this as a child to the active parent door.");
+			if (cwDoorCmds.parentData[door] != player.cwParentDoor) then
+				cwDoorCmds.parentData[door] = player.cwParentDoor;
+				cwDoorCmds:SaveParentData();		
+
+				Clockwork.entity:SetDoorParent(door, player.cwParentDoor);
+				Clockwork.player:Notify(player, "You have added this as a child to the active parent door.");
+
+				cwDoorCmds.infoTable = cwDoorCmds.infoTable or {};
+				table.insert(cwDoorCmds.infoTable, door)
+
+				Clockwork.datastream:Start(player, "doorParentESP", cwDoorCmds.infoTable);
+			else
+				Clockwork.player:Notify(player, "This door is already a child to the active parent door!");
+			end;
 		else
 			Clockwork.player:Notify(player, "You have not selected a valid parent door!");
 		end;